Scriptoria/scriptoria/requirements.txt
streamlit==1.31.0
requests==2.31.0
python-dotenv==1.0.0
huggingface-hub==0.20.3
Pillow==10.2.0
fpdf==1.7.2
